# docs/swagger.yaml
basePath: /
definitions:
  dtos.Campaign:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      masterId:
        type: string
      status:
        $ref: '#/definitions/dtos.CampaignStatus'
      system:
        $ref: '#/definitions/dtos.GameSystem'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  dtos.CampaignSessionTie:
    properties:
      briefDescription:
        type: string
      cachedScheduledAt:
        type: string
      cachedTitle:
        type: string
      campaignId:
        type: string
      orderIndex:
        type: integer
      sessionId:
        type: string
    type: object
  dtos.CampaignStatus:
    enum:
    - active
    - completed
    - cancelled
    - paused
    type: string
    x-enum-varnames:
    - CampaignActive
    - CampaignCompleted
    - CampaignCancelled
    - CampaignPaused
  dtos.Character:
    properties:
      avatarUrl:
        type: string
      class:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      level:
        type: integer
      name:
        type: string
      playerId:
        type: string
      sheetUrl:
        type: string
      updatedAt:
        type: string
    type: object
  dtos.GameSystem:
    properties:
      badgeColor:
        type: string
      id:
        type: string
      isCurated:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
...
  dtos.SessionPlayer:
    properties:
      character:
        $ref: '#/definitions/dtos.Character'
      joinedAt:
        type: string
      playerId:
        type: string
      sessionId:
        type: string
      status:
        $ref: '#/definitions/dtos.SessionPlayerStatus'
    type: object
  dtos.SessionPlayerStatus:
    enum:
    - active
    - kicked
    - left
    type: string
    x-enum-varnames:
    - PlayerActive
    - PlayerKicked
    - PlayerLeft
...
  internal_handlers.ChangeCampaignStatusRequest:
    properties:
      status:
        type: string
    type: object
...
  internal_handlers.CreateCampaignRequest:
    properties:
      description:
        type: string
      systemId:
        type: string
      title:
        type: string
    type: object
...
  internal_handlers.EditCampaignRequest:
    properties:
      description:
        type: string
      systemId:
        type: string
      title:
        type: string
    type: object
...
  internal_handlers.EditTieRequest:
    properties:
      briefDescription:
        type: string
      orderIndex:
        type: integer
    type: object
  internal_handlers.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_handlers.TieSessionRequest:
    properties:
      briefDescription:
        type: string
      orderIndex:
        type: integer
      sessionId:
        type: string
    type: object
host: localhost:5137
info:
  contact: {}
  description: Session, game system, campaign and character API for Session Zero
  title: Session Zero Session Service
  version: "1.0"
paths:
  /v1/campaigns:
    get:
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter by master user ID
        in: query
        name: masterId
        type: string
      - description: Filter by game system ID
        in: query
        name: systemId
        type: string
      - description: Status filter
        enum:
        - active
        - completed
        - cancelled
        - paused
        in: query
        name: status
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  $ref: '#/definitions/dtos.Campaign'
                type: array
              nextCursor:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: List campaigns
      tags:
      - campaigns
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Campaign'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create a campaign
      tags:
      - campaigns
  /v1/campaigns/{id}:
    delete:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete a campaign
      tags:
      - campaigns
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Campaign'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get campaign by ID
      tags:
      - campaigns
    patch:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.EditCampaignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Campaign'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Edit a campaign
      tags:
      - campaigns
  /v1/campaigns/{id}/players:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.SessionPlayer'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: List players in a campaign
      tags:
      - campaigns
  /v1/campaigns/{id}/sessions:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CampaignSessionTie'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: List sessions tied to a campaign
      tags:
      - campaigns
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Session tie data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.TieSessionRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Tie a session to a campaign
      tags:
      - campaigns
  /v1/campaigns/{id}/sessions/{sessionId}:
    delete:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Remove a session from a campaign
      tags:
      - campaigns
    patch:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Tie metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.EditTieRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update campaign-session tie metadata
      tags:
      - campaigns
  /v1/campaigns/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ChangeCampaignStatusRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Change campaign status
      tags:
      - campaigns
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"